move codicon to base/common
