+++ src/vs/base/browser/ui/findinput/findInputCheckboxes.ts
@@ -6,7 +6,7 @@
 import { Checkbox } from 'vs/base/browser/ui/checkbox/checkbox';
 import { Color } from 'vs/base/common/color';
 import * as nls from 'vs/nls';
-import { Codicon } from 'vs/base/browser/ui/codicons/codicons';
+import { Codicon } from 'vs/base/common/codicons';
 
 export interface IFindInputCheckboxOpts {
 	readonly appendTitle: string;
