+++ src/vs/base/browser/ui/scrollbar/verticalScrollbar.ts
@@ -9,7 +9,7 @@ import { ScrollableElementResolvedOptions } from 'vs/base/browser/ui/scrollbar/s
 import { ARROW_IMG_SIZE } from 'vs/base/browser/ui/scrollbar/scrollbarArrow';
 import { ScrollbarState } from 'vs/base/browser/ui/scrollbar/scrollbarState';
 import { INewScrollPosition, ScrollEvent, Scrollable, ScrollbarVisibility } from 'vs/base/common/scrollable';
-import { Codicon, registerIcon } from 'vs/base/browser/ui/codicons/codicons';
+import { Codicon, registerIcon } from 'vs/base/common/codicons';
 
 const scrollbarButtonUpIcon = registerIcon('scrollbar-button-up', Codicon.triangleUp);
 const scrollbarButtonDownIcon = registerIcon('scrollbar-button-down', Codicon.triangleDown);
