+++ src/vs/base/parts/quickinput/browser/quickInput.ts
@@ -27,7 +27,7 @@ import { IListVirtualDelegate, IListRenderer } from 'vs/base/browser/ui/list/lis
 import { List, IListOptions, IListStyles } from 'vs/base/browser/ui/list/listWidget';
 import { IInputBoxStyles } from 'vs/base/browser/ui/inputbox/inputBox';
 import { Color } from 'vs/base/common/color';
-import { registerIcon, Codicon } from 'vs/base/browser/ui/codicons/codicons';
+import { registerIcon, Codicon } from 'vs/base/common/codicons';
 
 export interface IQuickInputOptions {
 	idPrefix: string;
