+++ src/vs/editor/contrib/folding/foldingDecorations.ts
@@ -7,7 +7,7 @@ import { TrackedRangeStickiness, IModelDeltaDecoration, IModelDecorationsChangeA
 import { ModelDecorationOptions } from 'vs/editor/common/model/textModel';
 import { IDecorationProvider } from 'vs/editor/contrib/folding/foldingModel';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
-import { Codicon, registerIcon } from 'vs/base/browser/ui/codicons/codicons';
+import { Codicon, registerIcon } from 'vs/base/common/codicons';
 
 const foldingExpandedIcon = registerIcon('folding-expanded', Codicon.chevronDown);
 const foldingCollapsedIcon = registerIcon('folding-collapsed', Codicon.chevronRight);
